backend/requirements.txt
fastapi
pydantic
mysql-connector-python
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart